BOCM-358 - refactoring
