+++ apps/api/src/server/app/state-machine/lpa-questionnaire-states.js
@@ -1,4 +1,3 @@
-import mapObjectKeysToStrings from '../utils/map-states-to-strings.js';
 import lpaQuestionnaireActionsService from './lpa-questionnaire.actions.js';
 
 const lpaQuestionnaireActions = {
